Fixed Bridge Collider
Before making the build, please set the position of barriers in the oil rig plane as I have changed the position of the whole plane due to the bridge.

Also, all the bridge collider problems are fixed, if the tank stops due to some collider, make it as non-convex, it will work.

TRY PLAYING THE GAME BEFORE THE BUILD TO CHECK EVERYTHING